+++ antipetros_discordbot/cogs/community_events_cogs/give_away_cog.py
@@ -24,7 +24,7 @@ from antipetros_discordbot.utility.gidtools_functions import loadjson, pathmaker
 from antipetros_discordbot.init_userdata.user_data_setup import ParaStorageKeeper
 from antipetros_discordbot.utility.poor_mans_abc import attribute_checker
 from antipetros_discordbot.utility.emoji_handling import normalize_emoji
-from antipetros_discordbot.utility.replacements.command_replacement import auto_meta_info_command, AntiPetrosFlagCommand
+from antipetros_discordbot.engine.replacements import auto_meta_info_command, AntiPetrosFlagCommand
 from antipetros_discordbot.auxiliary_classes.for_cogs.aux_give_away_cog import GiveAwayEvent
 from antipetros_discordbot.utility.discord_markdown_helper.special_characters import ZERO_WIDTH
