+++ antipetros_discordbot/cogs/general_cogs/klim_bim_cog.py
@@ -31,7 +31,7 @@ from antipetros_discordbot.utility.discord_markdown_helper.the_dragon import THE
 from antipetros_discordbot.utility.discord_markdown_helper.special_characters import ZERO_WIDTH
 from antipetros_discordbot.utility.poor_mans_abc import attribute_checker
 from antipetros_discordbot.utility.enums import RequestStatus, CogState, UpdateTypus
-from antipetros_discordbot.utility.replacements.command_replacement import auto_meta_info_command
+from antipetros_discordbot.engine.replacements import auto_meta_info_command
 from antipetros_discordbot.utility.gidtools_functions import bytes2human
 from antipetros_discordbot.utility.exceptions import ParseDiceLineError
