+++ antipetros_discordbot/engine/replacements/__init__.py
@@ -1,0 +1,2 @@
+from .command_replacements import AntiPetrosBaseCommand, AntiPetrosFlagCommand, auto_meta_info_command
+from .help_replacement import BaseCustomHelpCommand
