+++ antipetros_discordbot/engine/replacements/command_replacements/__init__.py
@@ -1,0 +1,3 @@
+from .base_command import AntiPetrosBaseCommand
+from .flag_command import AntiPetrosFlagCommand
+from .creation_decorators import auto_meta_info_command
