+++ antipetros_discordbot/engine/replacements/command_replacements/creation_decorators.py
@@ -1,0 +1,169 @@
+"""
+[summary]
+
+[extended_summary]
+"""
+
+# region [Imports]
+
+# * Standard Library Imports ------------------------------------------------------------------------------------------------------------------------------------>
+
+import gc
+import os
+import re
+import sys
+import json
+import lzma
+import time
+import queue
+import base64
+import pickle
+import random
+import shelve
+import shutil
+import asyncio
+import logging
+import sqlite3
+import platform
+import importlib
+import subprocess
+import unicodedata
+
+from io import BytesIO
+from abc import ABC, abstractmethod
+from copy import copy, deepcopy
+from enum import Enum, Flag, auto
+from time import time, sleep
+from pprint import pprint, pformat
+from string import Formatter, digits, printable, whitespace, punctuation, ascii_letters, ascii_lowercase, ascii_uppercase
+from timeit import Timer
+from typing import Union, Callable, Iterable
+from inspect import stack, getdoc, getmodule, getsource, getmembers, getmodulename, getsourcefile, getfullargspec, getsourcelines
+from zipfile import ZipFile
+from datetime import tzinfo, datetime, timezone, timedelta
+from tempfile import TemporaryDirectory
+from textwrap import TextWrapper, fill, wrap, dedent, indent, shorten
+from functools import wraps, partial, lru_cache, singledispatch, total_ordering
+from importlib import import_module, invalidate_caches
+from contextlib import contextmanager
+from statistics import mean, mode, stdev, median, variance, pvariance, harmonic_mean, median_grouped
+from collections import Counter, ChainMap, deque, namedtuple, defaultdict
+from urllib.parse import urlparse
+from importlib.util import find_spec, module_from_spec, spec_from_file_location
+from concurrent.futures import ThreadPoolExecutor, ProcessPoolExecutor
+from importlib.machinery import SourceFileLoader
+
+
+# * Third Party Imports ----------------------------------------------------------------------------------------------------------------------------------------->
+
+import discord
+
+# import requests
+
+# import pyperclip
+
+# import matplotlib.pyplot as plt
+
+# from bs4 import BeautifulSoup
+
+# from dotenv import load_dotenv
+
+# from discord import Embed, File
+
+from discord.ext import commands, tasks
+
+# from github import Github, GithubException
+
+# from jinja2 import BaseLoader, Environment
+
+# from natsort import natsorted
+
+# from fuzzywuzzy import fuzz, process
+
+
+# * Gid Imports ------------------------------------------------------------------------------------------------------------------------------------------------->
+
+import gidlogger as glog
+
+from .base_command import AntiPetrosBaseCommand
+
+
+# endregion[Imports]
+
+# region [TODO]
+
+
+# endregion [TODO]
+
+# region [AppUserData]
+
+
+# endregion [AppUserData]
+
+# region [Logging]
+
+log = glog.aux_logger(__name__)
+log.info(glog.imported(__name__))
+
+# endregion[Logging]
+
+# region [Constants]
+
+THIS_FILE_DIR = os.path.abspath(os.path.dirname(__file__))
+
+# endregion[Constants]
+
+
+def auto_meta_info_command(name=None, cls=None, **attrs):
+    """
+    EXTENDED_BY_GIDDI
+    -----------------
+    Automatically gets the following attributes, if not provided or additional to provided:
+    - creates default aliases and retrieves custom aliases.
+
+    Base Docstring
+    ---------------
+    A decorator that transforms a function into a :class:`.Command`
+    or if called with :func:`.group`, :class:`.Group`.
+
+    By default the ``help`` attribute is received automatically from the
+    docstring of the function and is cleaned up with the use of
+    ``inspect.cleandoc``. If the docstring is ``bytes``, then it is decoded
+    into :class:`str` using utf-8 encoding.
+
+    All checks added using the :func:`.check` & co. decorators are added into
+    the function. There is no way to supply your own checks through this
+    decorator.
+
+    Parameters
+    -----------
+    name: :class:`str`
+        The name to create the command with. By default this uses the
+        function name unchanged.
+    cls
+        The class to construct with. By default this is :class:`.Command`.
+        You usually do not change this.
+    attrs
+        Keyword arguments to pass into the construction of the class denoted
+        by ``cls``.
+
+    Raises
+    -------
+    TypeError
+        If the function is not a coroutine or is already a command.
+    """
+    if cls is None:
+        cls = AntiPetrosBaseCommand
+
+    def decorator(func):
+        return cls(func, name=name, **attrs)
+
+    return decorator
+
+
+# region[Main_Exec]
+
+if __name__ == '__main__':
+    pass
+
+# endregion[Main_Exec]
