+++ antipetros_discordbot/engine/replacements/command_replacements/flag_command.py
@@ -1,0 +1,132 @@
+"""
+[summary]
+
+[extended_summary]
+"""
+
+# region [Imports]
+
+# * Standard Library Imports ------------------------------------------------------------------------------------------------------------------------------------>
+
+import gc
+import os
+import re
+import sys
+import json
+import lzma
+import time
+import queue
+import base64
+import pickle
+import random
+import shelve
+import shutil
+import asyncio
+import logging
+import sqlite3
+import platform
+import importlib
+import subprocess
+import unicodedata
+
+from io import BytesIO
+from abc import ABC, abstractmethod
+from copy import copy, deepcopy
+from enum import Enum, Flag, auto
+from time import time, sleep
+from pprint import pprint, pformat
+from string import Formatter, digits, printable, whitespace, punctuation, ascii_letters, ascii_lowercase, ascii_uppercase
+from timeit import Timer
+from typing import Union, Callable, Iterable
+from inspect import stack, getdoc, getmodule, getsource, getmembers, getmodulename, getsourcefile, getfullargspec, getsourcelines
+from zipfile import ZipFile
+from datetime import tzinfo, datetime, timezone, timedelta
+from tempfile import TemporaryDirectory
+from textwrap import TextWrapper, fill, wrap, dedent, indent, shorten
+from functools import wraps, partial, lru_cache, singledispatch, total_ordering
+from importlib import import_module, invalidate_caches
+from contextlib import contextmanager
+from statistics import mean, mode, stdev, median, variance, pvariance, harmonic_mean, median_grouped
+from collections import Counter, ChainMap, deque, namedtuple, defaultdict
+from urllib.parse import urlparse
+from importlib.util import find_spec, module_from_spec, spec_from_file_location
+from concurrent.futures import ThreadPoolExecutor, ProcessPoolExecutor
+from importlib.machinery import SourceFileLoader
+import inspect
+
+# * Third Party Imports ----------------------------------------------------------------------------------------------------------------------------------------->
+
+import discord
+
+# import requests
+
+# import pyperclip
+
+# import matplotlib.pyplot as plt
+
+# from bs4 import BeautifulSoup
+
+# from dotenv import load_dotenv
+
+# from discord import Embed, File
+
+from discord.ext import commands, tasks, flags, ipc
+
+# from github import Github, GithubException
+
+# from jinja2 import BaseLoader, Environment
+
+# from natsort import natsorted
+
+# from fuzzywuzzy import fuzz, process
+
+
+# * Gid Imports ------------------------------------------------------------------------------------------------------------------------------------------------->
+
+import gidlogger as glog
+
+
+# * Local Imports ----------------------------------------------------------------------------------------------------------------------------------------------->
+
+from antipetros_discordbot.utility.gidtools_functions import loadjson, writejson, pathmaker
+from antipetros_discordbot.init_userdata.user_data_setup import ParaStorageKeeper
+from .helper.alias_provider import AliasProvider
+from .base_command import AntiPetrosBaseCommand
+# endregion[Imports]
+
+# region [TODO]
+
+
+# endregion [TODO]
+
+# region [AppUserData]
+
+APPDATA = ParaStorageKeeper.get_appdata()
+BASE_CONFIG = ParaStorageKeeper.get_config('base_config')
+
+
+# endregion [AppUserData]
+
+# region [Logging]
+
+log = glog.aux_logger(__name__)
+log.info(glog.imported(__name__))
+
+# endregion[Logging]
+
+# region [Constants]
+
+THIS_FILE_DIR = os.path.abspath(os.path.dirname(__file__))
+
+# endregion[Constants]
+
+
+class AntiPetrosFlagCommand(flags.FlagCommand, AntiPetrosBaseCommand):
+    pass
+
+
+# region[Main_Exec]
+if __name__ == '__main__':
+    pass
+
+# endregion[Main_Exec]
