+++ antipetros_discordbot/engine/replacements/help_replacement.py
@@ -1,7 +1,7 @@
 from discord.ext.commands import Command
-from antipetros_discordbot.utility.gidtools_functions import loadjson
-from antipetros_discordbot.init_userdata.user_data_setup import ParaStorageKeeper
 from antipetros_discordbot.utility.gidtools_functions import pathmaker, readit, writeit, loadjson, writejson
+from antipetros_discordbot.init_userdata.user_data_setup import ParaStorageKeeper
+
 from typing import List, Dict, Set, Tuple, Union
 import os
 import discord
