--- antipetros_discordbot/utility/replacements/__init__.py
@@ -1,1 +1,0 @@
-from .command_replacement import auto_meta_info_command
